Single responsibillity principle
Separate the clases by use
