+++ photo-app/src/item.ts
@@ -1,0 +1,5 @@
+export class Item {
+    constructor(public readonly id: number,
+                protected title: string) {
+    }
+}
