+++ photo-app/src/photo-orientation.ts
@@ -1,0 +1,6 @@
+export enum PhotoOrientation {
+    Landscape,
+    Portrait,
+    Square,
+    Panorama
+}
